[kotlin] Fix test BuiltInsReferenceResolverTest#testAllReferencesResolved
So... this test actually doing the following:
- loading the built-ins from the class path
- for every declaration from built ins it is trying to navigate to the
sources
- sources here are provided by KotlinArtifacts

This test was broken since the kotlin/1564f2c549bbdff361c02f71aa128e06274336ce
That had happened by two reasons.
1) in the mentioned commit some methods were annotated with the new
annotation (IntrinsicConstEvaluation) with runtime retention.
Because this annotation has the runtime retention, it became a part of the
key in the ByDescriptorIndexer.
2) `intellij.platform.core` also contains the stdlib but some old version bundled
into the platform. So we are loading the built-ins with an old version without
IntrinsicConstEvaluation and then trying to navigate to the sources with this annotation.

So this commit are fixing the reason number 2 -- so we will use the built-ins with the same version.
That is quite logical btw, because some methods could be legitimately changed in the built-ins
so the same version should be used in the first place.

Also, it is a bit strange, that we are using platform stdlib in the tests, because on runtime
we are actually having our own stdlib inside the kotlin plugin at the moment.

This test are using builtIns loaded from classpath and trying to navigate
to the source code. This test are using KotlinArtifacts to set up the project
with the sources, so it will use the stdlib with the version specified
in .libraries/kotlinc_kotlin_dist.xml.

But `intellij.platform.core` also contains the stdlib,
but some old version. And primitives were recently changed
in the kotlin/1564f2c549bbdff361c02f71aa128e06274336ce --
binary annotation was added (IntrinsicConstEvaluation)

GitOrigin-RevId: 912869a7f7d9ba98858022be7faf38518ec6dca2